Reduce copying inside SourceFile::setContents; fix non-BOM UTF-16 offset and skip it if file size is odd
Previously, `SourceFioe::setContents(ISlangBlob*)` would always copy the
contents twice: once inside `CharEncoding::decode`, and once when creating
a `RawBlob`.

This commit adds a fast-path for the common case of ASCII or UTF-8 without
a BOM, in which case we don't need to copy at all.
In the slow path we can also save a copy by moving the list's buffer into the
RawBlob (though maybe there's a more ergonomic way to do this?).

Note that this isn't quite optimal for `SourceFile::setContents(const String&)`
when the input is UTF-16. The `const String&` version always copies at least once,
which is unnecessary when we need to copy on the decode.

This also includes a quick fix for `CharEncoding::determineEncoding`: if it
determines the file is UTF-16 without a BOM, it reported an offset of 2 when
it should have reported an offset of 0. Additionally, we can skip the nontrivial
(5 instructions/byte) search for UTF-16 patterns half the time; if the file
length is odd, it can't be UTF-16.

Signed-off-by: Nia Bickford <nbickford@nvidia.com>